♻️ refactor : 디렉토리 구조 총 변경
과제 제출 과정에서 빌드중에 문제가 있는 듯 보임.
따라서 추가했던 폴더를 삭제하고 다시 src안으로 모든 파일을 모음.
