--- src/constants/index.js
@@ -1,3 +1,0 @@
-export { default as MAGICNUMBER } from './magicnumber.js';
-export { default as ERRORMESAGE } from './errorMessage.js';
-export { default as IOMESSAGE } from './ioMessage.js';
