+++ src/magicnumber.js
