--- src/utils/index.js
@@ -1,3 +1,0 @@
-export { default as ioMethod } from './inputoutputMethod.js';
-export { default as validation } from './validation.js';
-export { default as pickRandomNumber } from './pickRandomNumber.js';
